Abstractize to add database implementation
